### 2026-01-17 #9806 @arfon
### Update regarding JOSS scope and paper format changes

:wave: @shin1koda. Thank you for your patience while your submission has been waiting in our pre-review queue.

#### What this means for your submission

For full details, please see:

- Our updated scope definition: https://joss.readthedocs.io/en/latest/submitting.html#scope-and-significance
- An example paper in the new format: https://joss.readthedocs.io/en/latest/example_paper.html

**Important:** Once you have updated your paper, please comment here to let us know. At that point, we will assess your software against our updated scope criteria. We want to be transparent that under these new criteria, some submissions that were previously in scope may now be deemed out of scope, and we may choose not to proceed with review.
