### 2026-01-18 71d60f4
modify prallel calculation
