+++ README.md
@@ -96,10 +96,10 @@ Running the script produces several output files, including:
 `torch_dmf` is a PyTorch-accelerated backend that mirrors the `dmf` API while offloading internal tensor operations to PyTorch (CUDA when available).
 
 If you want to use `torch_dmf`, install a CUDA-matched build of PyTorch before installing PyDMF.  
-Example for CUDA 12.9:
+Example for CUDA 12.8:
 
 ```bash
-pip install torch --index-url https://download.pytorch.org/whl/cu129
+pip install torch --index-url https://download.pytorch.org/whl/cu128
 pip install git+https://github.com/t-0hmura/dmf.git
 ```
 
