### 2026-01-23 1fb3632
add guard for t_eval update
