+++ src/torch_dmf/dmf.py
@@ -150,7 +150,8 @@ class VariationalPathOpt(ABC, cyipopt.Problem):
 
     remove_rotation_and_translation : bool, optional
         If True, remove global translational and rotational motion using
-        nonlinear constraints. Default: True.
+        nonlinear constraints. Default: True. Automatically disabled when
+        atomic constraints are present on ``ref_images``.
 
     mass_weighted : bool, optional
         If True, the velocity norm \( \vert \dot{x}(t) \vert \) uses mass-weighted
@@ -298,8 +299,12 @@ class VariationalPathOpt(ABC, cyipopt.Problem):
         self._mass_fracs = self._masses/np.sum(self._masses)
 
         #Constraints
-        self.remove_rotation_and_translation \
-            = remove_rotation_and_translation
+        has_constraints = any(
+            getattr(image, "constraints", None) for image in ref_images
+        )
+        if has_constraints:
+            remove_rotation_and_translation = False
+        self.remove_rotation_and_translation = remove_rotation_and_translation
         self.eps_vel = float(eps_vel)
         self.eps_rot = float(eps_rot)
 
@@ -435,6 +440,10 @@ class VariationalPathOpt(ABC, cyipopt.Problem):
             'limited_memory_init_val':2.5,
             'accept_every_trial_step':'yes',
             'output_file':'pathopt.out',
+
+            "mumps_mem_percent": 2,
+            "hessian_approximation": "limited-memory",
+            "limited_memory_max_history": 5,
             }
 
         if self.parallel and self._world.size>1:
@@ -690,25 +699,30 @@ class VariationalPathOpt(ABC, cyipopt.Problem):
                 self._P_rot[0,:,1:],
                 y[:-1])
         return jac_rot
-
     def _get_consts_vel(self) -> np.ndarray:
         pos_t = self._get_positions_torch(self._P_vel_t)
-        diffs = pos_t[1:]-pos_t[:-1]
-        d2s = torch.sum(self._masses_t[None,:,None]*diffs**2,dim=(1,2))
-        return (d2s/torch.mean(d2s)).cpu().numpy()
-
+        diffs = pos_t[1:] - pos_t[:-1]
+        d2s = torch.sum(self._masses_t[None, :, None] * diffs**2, dim=(1, 2))
+        ave = torch.mean(d2s)
+        # Guard against division-by-zero in pathological trial steps
+        ave = torch.clamp(ave, min=1.0e-300)
+        return (d2s / ave).cpu().numpy()
     def _get_jac_vel(self) -> np.ndarray:
         pos_t = self._get_positions_torch(self._P_vel_t)
-        diffs = pos_t[1:]-pos_t[:-1]
-        d2s = torch.sum(self._masses_t[None,:,None]*diffs**2,dim=(1,2))
-        diff_P = self._P_vel_t[0,:,1:]-self._P_vel_t[0,:,:-1]
-        jac_d2s = 2.0*torch.einsum(
+        diffs = pos_t[1:] - pos_t[:-1]
+        d2s = torch.sum(self._masses_t[None, :, None] * diffs**2, dim=(1, 2))
+        diff_P = self._P_vel_t[0, :, 1:] - self._P_vel_t[0, :, :-1]
+        jac_d2s = 2.0 * torch.einsum(
             'a,bi,ias->ibas',
-            self._masses_t,diff_P,diffs)
+            self._masses_t, diff_P, diffs
+        )
         ave_d2s = torch.mean(d2s)
-        return (jac_d2s/ave_d2s \
-            - torch.tensordot(d2s,torch.mean(jac_d2s,dim=0),0)/(ave_d2s)**2
-            ).cpu().numpy()
[…]
