+++ src/torch_dmf_/fbenm.py
@@ -52,16 +52,31 @@ class FB_ENM(Calculator):
 
     Parameters
     ----------
-    d_min, d_max : array-like, shape (N, N)
-        Flat-bottom bounds for each pair distance.
-    delta_min, delta_max : array-like, shape (N, N) or None
-        Quadratic wall widths. If None, generated via ``delta_scale``.
+    d_min : ndarray of shape (N, N)
+        Lower bounds of the flat-bottom region for each atom pair.
+        Here, N must match the number of atoms in the Atoms object
+        to which this calculator is attached.
+    d_max : ndarray of shape (N, N)
+        Upper bounds of the flat-bottom region for each atom pair.
+        Must have the same shape and atom ordering as d_min.
+    delta_min, delta_max : ndarray of shape (N, N) or None
+        Width parameters for the quadratic walls. If None, they are
+        set to ``delta_scale * d_min`` and ``delta_scale * d_max``,
+        respectively. Default: None.
     delta_scale : float
-        Scale factor used when delta arrays are autogenerated.
-    return_energy_mats : bool
-        If True, include ``emat_rep`` and ``emat_att`` in results.
+        Scale factor used when delta_min/max are autogenerated.
+        Default: 0.2.
+    return_energy_mats : bool, optional
+        If False, omit energy matrices from results. Default: True.
     device : str or torch.device, optional
-        Torch device used only for the distance-matrix evaluation.
+        Torch device for internal tensors. If None, auto-select.
+
+    Notes
+    -----
+    - Energy is zero when all distances lie within [d_min, d_max].
+    - Forces derive from quadratic penalties outside the flat bottom.
+    - This class does not determine d_min/d_max from structures;
+      see `FB_ENM_Bonds` for the bond-aware construction.
 
     """
 
@@ -77,87 +92,86 @@ class FB_ENM(Calculator):
         return_energy_mats: bool = False,
         device=None,
     ):
-        super().__init__()
+        Calculator.__init__(self)
 
         self.device = _resolve_torch_device(device)
+        _dev = self.device
+        _tdt = torch.float64
         self._return_energy_mats = bool(return_energy_mats)
 
-        def to_np(x) -> np.ndarray:
-            if isinstance(x, torch.Tensor):
-                return x.detach().cpu().numpy().astype(np.float64, copy=False)
-            return np.asarray(x, dtype=np.float64)
+        def to_t(x):
+            return torch.as_tensor(x,dtype=_tdt,device=_dev)
 
-        d_min_np = to_np(d_min).copy()
-        d_max_np = to_np(d_max).copy()
-
+        d_min_t = to_t(d_min).clone()
+        d_max_t = to_t(d_max).clone()
         if delta_min is not None:
-            delta_min_np = to_np(delta_min).copy()
+            delta_min_t = to_t(delta_min).clone()
         else:
-            delta_min_np = (float(delta_scale) * d_min_np).copy()
-
+            delta_min_t = (float(delta_scale)*d_min_t).clone()
         if delta_max is not None:
-            delta_max_np = to_np(delta_max).copy()
+            delta_max_t = to_t(delta_max).clone()
         else:
-            delta_max_np = (float(delta_scale) * d_max_np).copy()
+            delta_max_t = (float(delta_scale)*d_max_t).clone()
 
[…]
