+++ src/torch_dmf_/dmf.py
@@ -1,1901 +1,0 @@
-import threading
-from abc import ABC, abstractmethod
-from typing import Optional
-
-import numpy as np
-from numpy.polynomial import polynomial as P
-from scipy.interpolate import BSpline,interp1d
-from scipy.spatial.transform import Rotation
-import cyipopt
-
-import ase.parallel
-from functools import cached_property
-
-import torch
-import warnings
-
-
-def _resolve_torch_device(device_spec):
-    """
-    Resolve a torch device from a user spec.
-    """
-    if device_spec is None:
-        d = torch.device("cuda" if torch.cuda.is_available() else "cpu")
-    else:
-        try:
-            d = torch.device(device_spec)
-        except (TypeError, ValueError):
-            warnings.warn(f"[torch_dmf] Invalid device spec '{device_spec}', falling back to CPU.")
-            d = torch.device("cpu")
-    if d.type == "cuda" and not torch.cuda.is_available():
-        warnings.warn("[torch_dmf] CUDA requested but not available; falling back to CPU.")
-        d = torch.device("cpu")
-    return d
-
-
-@torch.no_grad()
-def _interp1d_torch(xq: torch.Tensor, xp: torch.Tensor, fp: torch.Tensor) -> torch.Tensor:
-    """
-    Linear interpolate fp(xp) at xq using pure PyTorch.
-
-    xp: (M,), fp: (M, *), xq: (K,)
-    Returns: (K, *) broadcast along the trailing dims of fp.
-    """
-    idx = torch.searchsorted(xp,xq,right=False).clamp(1,xp.numel()-1)
-    x0, x1 = xp[idx-1], xp[idx]
-    lam = (xq-x0)/(x1-x0)
-    while lam.dim() < fp.dim():
-        lam = lam.unsqueeze(-1)
-    f0, f1 = fp[idx-1], fp[idx]
-    return f0 + (f1-f0)*lam
-
-
-class HistoryBase():
-    """
-    Container storing the optimization history of the VariationalPathOpt.
-
-    This object collects various physical and numerical quantities evaluated
-    along the reaction path during the optimization.  At each IPOPT iteration,
-    the ``VariationalPathOpt.intermediate`` method appends the current values
-    of these quantities to the corresponding lists below.
-
-    Attributes
-    ----------
-    forces : list of ndarray
-        History of ``VariationalPathOpt.forces``.
-    energies : list of ndarray
-        History of ``VariationalPathOpt.energies``.
-    coefs : list of ndarray
-        History of ``VariationalPathOpt.coefs``.
-    angs : list of ndarray
-        History of ``VariationalPathOpt.angs``.
-    tmax : list of float
-        History of the location ``t_max`` corresponding to the maximum
-        interpolated energy along the path. See Ref. 1 for details.
-    images_tmax : list of ase.Atoms
-        History of the atomic structure at ``t = t_max``, providing an
-        approximate transition-state geometry at each iteration.
-    duals : list of float
-        History of the scaled dual infeasibility (IPOPT diagnostic).
-
-    """
-
-    def __init__(self):
-        self.forces = []
-        self.energies = []
-        self.coefs = []
-        self.angs = []
-        self.tmax = []
-        self.images_tmax = []
-        self.duals = []
-
-
-class VariationalPathOpt(ABC, cyipopt.Problem):
-    r"""
-    Abstract base class for variational reaction–path optimization.
-
-    This class formulates a general functional
-
-    .. math::
-
-        \tilde{I}[x(t)] = K(I[x(t)]),
-
-    where
-
-    .. math::
-
-        I[x(t)] = \int_0^1 dt\, \vert \dot{x}(t) \vert \, F(x(t)).
-
-    The functions \(K(I)\), \(F(x)\), and their derivatives are supplied
[…]
