+++ src/torch_dmf_/fbenm.py
@@ -1,728 +1,0 @@
-from typing import Optional, Union
-
-import numpy as np
-
-from ase.calculators.calculator import Calculator
-from ase.data import covalent_radii
-from ase.data.vdw_alvarez import vdw_radii
-
-import torch
-import warnings
-
-
-def _resolve_torch_device(device_spec):
-    """
-    Resolve a torch device from a user spec.
-    """
-    if device_spec is None:
-        d = torch.device("cuda" if torch.cuda.is_available() else "cpu")
-    else:
-        try:
-            d = torch.device(device_spec)
-        except (TypeError, ValueError):
-            warnings.warn(f"[torch_dmf] Invalid device spec '{device_spec}', falling back to CPU.")
-            d = torch.device("cpu")
-    if d.type == "cuda" and not torch.cuda.is_available():
-        warnings.warn("[torch_dmf] CUDA requested but not available; falling back to CPU.")
-        d = torch.device("cpu")
-    return d
-
-
-@torch.no_grad()
-def _calc_dist_mat(pos, *, device) -> torch.Tensor:
-    pos_t = torch.as_tensor(pos,dtype=torch.float64,device=device)
-    return torch.cdist(pos_t,pos_t)
-
-
-class FB_ENM(Calculator):
-    """
-    ASE calculator for the Flat-bottom Elastic Network Model (FB-ENM).
-
-    A lightweight structure-based potential for generating
-    collision-free plausible paths used in combination with the
-    direct MaxFlux method. Each pair interaction has a flat
-    bottom defined by (d_min, d_max). This class implements the
-    general pairwise model.
-
-    This implementation follows the model introduced in:
-
-        Koda & Saito, J. Chem. Theory Comput. 2024, 20, 7176–7187.
-
-    See the original paper for theoretical details.
-
-    Parameters
-    ----------
-    d_min : ndarray of shape (N, N)
-        Lower bounds of the flat-bottom region for each atom pair.
-        Here, N must match the number of atoms in the Atoms object
-        to which this calculator is attached.
-    d_max : ndarray of shape (N, N)
-        Upper bounds of the flat-bottom region for each atom pair.
-        Must have the same shape and atom ordering as d_min.
-    delta_min, delta_max : ndarray of shape (N, N) or None
-        Width parameters for the quadratic walls. If None, they are
-        set to ``delta_scale * d_min`` and ``delta_scale * d_max``,
-        respectively. Default: None.
-    delta_scale : float
-        Scale factor used when delta_min/max are autogenerated.
-        Default: 0.2.
-    return_energy_mats : bool, optional
-        If False, omit energy matrices from results. Default: True.
-    device : str or torch.device, optional
-        Torch device for internal tensors. If None, auto-select.
-
-    Notes
-    -----
-    - Energy is zero when all distances lie within [d_min, d_max].
-    - Forces derive from quadratic penalties outside the flat bottom.
-    - This class does not determine d_min/d_max from structures;
-      see `FB_ENM_Bonds` for the bond-aware construction.
-
-    """
-
-    implemented_properties = ['energy', 'forces']
-
-    def __init__(
-        self,
-        d_min,
-        d_max,
-        delta_min: Optional[Union[np.ndarray, torch.Tensor]] = None,
-        delta_max: Optional[Union[np.ndarray, torch.Tensor]] = None,
-        delta_scale: float = 0.2,
-        return_energy_mats: bool = False,
-        device=None,
-    ):
-        Calculator.__init__(self)
-
-        self.device = _resolve_torch_device(device)
[…]

[…]
