+++ paper/paper.bib
@@ -100,8 +100,16 @@
 @misc{pydmf,
   author = {Koda, Shin-ichi},
   title = {PyDMF: A Python package for double-ended reaction path/transition state optimization},
-  year = {2025},
+  year = {2026},
   publisher = {GitHub},
   journal = {GitHub repository},
   url = {https://github.com/shin1koda/dmf}
+}
+
+@misc{cyipopt,
+  title = {Cython interface for the interior point optimzer IPOPT},
+  year = {2026},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/mechmotum/cyipopt}
 }
