### 2026-01-30 fbe55a3
add some references to JOSS paper
