+++ paper/paper.bib
@@ -106,10 +106,24 @@
   url = {https://github.com/shin1koda/dmf}
 }
 
-@misc{cyipopt,
-  title = {Cython interface for the interior point optimzer IPOPT},
-  year = {2026},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  url = {https://github.com/mechmotum/cyipopt}
+@article{larsen2017theatomi,
+    author = {Larsen, Ask Hjorth and Mortensen, Jens J{\o}rgen and Blomqvist, Jakob and Castelli, Ivano E and Christensen, Rune and Du{\l}ak, Marcin and Friis, Jesper and Groves, Michael N and Hammer, Bj{\o}rk and Hargus, Cory and Hermes, Eric D and Jennings, Paul C and Jensen, Peter Bjerre and Kermode, James and Kitchin, John R and Kolsbjerg, Esben Leonhard and Kubal, Joseph and Kaasbjerg, Kristen and Lysgaard, Steen and Maronsson, J\'{o}n Bergmann and Maxson, Tristan and Olsen, Thomas and Pastewka, Lars and Peterson, Andrew and Rostgaard, Carsten and Schi{\o}tz, Jakob and Sch\"utt, Ole and Strange, Mikkel and Thygesen, Kristian S and Vegge, Tejs and Vilhelmsen, Lasse and Walter, Michael and Zeng, Zhenhua and Jacobsen, Karsten W},
+    year = "2017",
+    title = "{The atomic simulation environment---a Python library for working with atoms}",
+    journal = "Journal of Physics: Condensed Matter",
+    doi = "10.1088/1361-648x/aa680e",
+    pages = "273002",
+    number = "27",
+    volume = "29"
+}
+
+@article{wachter2006ontheimp,
+    author = {W\"{a}chter, Andreas and Biegler, Lorenz T.},
+    year = "2006",
+    title = "{On the implementation of an interior-point filter line-search algorithm for large-scale nonlinear programming}",
+    journal = "Mathematical Programming",
+    doi = "10.1007/s10107-004-0559-y",
+    pages = "25--57",
+    number = "1",
+    volume = "106"
 }
