+++ paper/paper.md
@@ -43,9 +43,9 @@ PyDMF provides Python implementations of both methods [@pydmf]. Because reaction
 
 # State of the field
 
-As noted above, existing double-ended optimization methods such as NEB are implemented internally in many atomistic simulation programs. The Atomic Simulation Environment (ASE), a Python package that provides a unified interface to many atomistic simulation programs, also provides its own implementations of NEB and several of its variants. In addition, the image-dependent pair potential (IDPP) method [@smidstrup2014improved] for generating initial paths is available in software such as ORCA and ASE.
+As noted above, existing double-ended optimization methods such as NEB are implemented internally in many atomistic simulation programs. The Atomic Simulation Environment (ASE) [@larsen2017theatomi], a Python package that provides a unified interface to many atomistic simulation programs, also provides its own implementations of NEB and several of its variants. In addition, the image-dependent pair potential (IDPP) method [@smidstrup2014improved] for generating initial paths is available in software such as ORCA and ASE.
 
-From a theoretical standpoint, DMF implemented in PyDMF differs fundamentally from the existing approaches. DMF is based on a variational formulation in which reaction-path optimization is expressed as a well-defined minimization problem with an explicit objective function. In contrast to NEB or string methods, which rely on non-variational schemes, DMF can directly leverage general-purpose nonlinear optimization algorithms. In practice, PyDMF employs the state-of-the-art optimizer IPOPT via its Python interface cyipopt, enabling efficient and robust optimization without introducing method-specific path-update algorithms.
+From a theoretical standpoint, DMF implemented in PyDMF differs fundamentally from the existing approaches. DMF is based on a variational formulation in which reaction-path optimization is expressed as a well-defined minimization problem with an explicit objective function. In contrast to NEB or string methods, which rely on non-variational schemes, DMF can directly leverage general-purpose nonlinear optimization algorithms. In practice, PyDMF employs the state-of-the-art optimizer IPOPT [@wachter2006ontheimp] via its Python interface cyipopt, enabling efficient and robust optimization without introducing method-specific path-update algorithms.
 
 This design choice necessarily introduces a dependency on external optimization libraries. For this reason, PyDMF was developed as a standalone software package rather than as a contribution to an existing framework. Requiring DMF-specific optimizer dependencies, which also require the use of conda, for all users of an integrated framework such as ASE would impose unnecessary constraints on users who do not require DMF functionality. Implementing PyDMF as an independent package therefore provides a practical separation of concerns, allowing advanced variational reaction-path optimization capabilities to be offered exclusively to users who need them.
 
