+++ src/dmf/dmf.py
@@ -1,5 +1,6 @@
 import threading
 from abc import ABC, abstractmethod
+from collections import defaultdict
 
 import numpy as np
 from numpy.polynomial import polynomial as P
@@ -7,9 +8,16 @@ from scipy.interpolate import BSpline,interp1d
 from scipy.spatial.transform import Rotation
 import cyipopt
 
-import ase.parallel
 from functools import cached_property
 
+from .mpi_backend import _worker_loop
+try:
+    import os
+    os.environ['UCX_LOG_LEVEL'] = 'error'
+    from mpi4py import MPI
+    HAS_MPI4PY = True
+except ImportError:
+    HAS_MPI4PY = False
 
 class HistoryBase():
     """
@@ -111,8 +119,7 @@ class VariationalPathOpt(ABC, cyipopt.Problem):
 
     remove_rotation_and_translation : bool, optional
         If True, remove global translational and rotational motion using
-        nonlinear constraints. Default: True. Automatically disabled when
-        atomic constraints are present on ``ref_images``.
+        nonlinear constraints. Default: True.
 
     mass_weighted : bool, optional
         If True, the velocity norm \( \vert \dot{x}(t) \vert \) uses mass-weighted
@@ -123,8 +130,7 @@ class VariationalPathOpt(ABC, cyipopt.Problem):
         Default: False.
 
     world : MPI communicator, optional
-        Communicator used when ``parallel=True``. Defaults to
-        ``ase.parallel.world``.
+        Communicator used when ``parallel=True``. Default: None.
 
     t_eval : ndarray, optional
         Energy evaluation points in \( t \in [0,1] \).  
@@ -240,11 +246,69 @@ class VariationalPathOpt(ABC, cyipopt.Problem):
         coefs=None, nsegs=4,dspl=3,
         remove_rotation_and_translation=True,
         mass_weighted=False,
+        calc_factory=None,
         parallel=False,world=None,
         t_eval=None,w_eval=None,
         n_vel=None,n_trans=None,n_rot=None,
         eps_vel=0.01,eps_rot=0.01,
         ):
+
+        #Prallel calculation
+        self.parallel = parallel
+
+        if parallel and (parallel == "mpi" or world is not None):
+            self.parallel = "mpi"
+            if not HAS_MPI4PY:
+                raise RuntimeError("MPI parallel calculation requires mpi4py.")
+
+            self._world = MPI.COMM_WORLD if world is None else world
+            self._rank = self._world.Get_rank()
+            self._size = self._world.Get_size()
+        else:
+            self._world = None
+            self._rank = 0
+            self._size = 1
+
+
+        #Initialize images
+        if t_eval is None:
+            self._nimages = 2*nsegs+1
+        else:
+            self._nimages = len(t_eval)
+
+        self.images=[]
+        for _ in range(self._nimages):
+            self.images.append(ref_images[0].copy())
+
+        r2i = defaultdict(list)
+        for i in range(self._nimages):
+            if i==0:
+                r = 0
+            elif i==self._nimages-1:
+                r = self._size-1
+            else:
+                r = (i-1)%self._size
+
+            r2i[r].append(i)
+
+        self._r2i = r2i
+
+        #calc_factory
+        self.calc_factory = calc_factory
+
+        if self.parallel == "mpi":
+            if self.calc_factory is None:
+                raise RuntimeError("parallel='mpi' requires calc_factory.")
+
+        if self.calc_factory is not None:
+            for i in self._r2i[self._rank]:
+                self.images[i].calc = self.calc_factory(i)
+
+        if self.parallel == "mpi" and self._rank>0:
+            _worker_loop(self._world, self.images)
+            import sys
+            sys.exit(0)
+
 
         #Atoms
         self.natoms = len(ref_images[0])
[…]
