### 2026-02-12 90cdfbd
remove np.float32
