+++ src/dmf/torch/fbenm.py
@@ -152,6 +152,17 @@ class FB_ENM(Calculator):
         self.delta_min = delta_min_np
         self.delta_max = delta_max_np
 
+    def copy(self):
+        return FB_ENM(
+            self.d_min,
+            self.d_max,
+            delta_min=self.delta_min,
+            delta_max=self.delta_max,
+            return_energy_mats=self._return_energy_mats,
+            device=self.device,
+            dtype=self.torch_dtype,
+        )
+
     @torch.no_grad()
     def calculate(self, atoms, properties, system_changes):
         super().calculate(atoms, properties, system_changes)
@@ -331,7 +342,7 @@ class FB_ENM_Bonds(FB_ENM):
                 J = (d/r_cov)<bond_scale_t
                 J_t = J.to(_tdt)
                 A = (J_t@J_t)>0
-                del J, J_t; torch.cuda.empty_cache()
+                del J, J_t
 
                 if fix_planes:
                     A = A|addA_p
@@ -342,10 +353,10 @@ class FB_ENM_Bonds(FB_ENM):
 
                 cand_min = torch.where(A,d,torch.minimum(d,r_vdw))
                 cand_max = torch.where(A,d,2.0*torch.amax(d))
-                del d, A; torch.cuda.empty_cache()
+                del d, A
                 d_min = torch.minimum(d_min,cand_min)
                 d_max = torch.maximum(d_max,cand_max)
-                del cand_min, cand_max; torch.cuda.empty_cache()
+                del cand_min, cand_max
 
             if (d_min_overwrite is not None) and (A_overwrite is not None):
                 mask = torch.as_tensor(A_overwrite,dtype=torch.bool,device=_dev)
@@ -354,9 +365,9 @@ class FB_ENM_Bonds(FB_ENM):
                 mask = torch.as_tensor(A_overwrite,dtype=torch.bool,device=_dev)
                 d_max = torch.where(mask,torch.as_tensor(d_max_overwrite,dtype=_tdt,device=_dev),d_max)
 
-        del numbers, r_cov_np, r_vdw_np, r_cov, r_vdw, bond_scale_t, natoms, addA_p, addA_mask, delA_mask; torch.cuda.empty_cache()
+        del numbers, r_cov_np, r_vdw_np, r_cov, r_vdw, bond_scale_t, natoms, addA_p, addA_mask, delA_mask
         super().__init__(d_min, d_max, delta_scale=delta_scale, device=device, dtype=self.torch_dtype)
-        del d_min, d_max; torch.cuda.empty_cache()
+        del d_min, d_max
 
 
 class CFB_ENM(Calculator):
@@ -436,7 +447,8 @@ class CFB_ENM(Calculator):
 
     def __init__(
         self,
-        images,
+        images=None,
+        d_bond: Optional[np.ndarray] = None,
         bond_scale: float = 1.25,
         d_corr0: Optional[np.ndarray] = None,
         corr0_scale: float = 1.10,
@@ -445,6 +457,7 @@ class CFB_ENM(Calculator):
         d_corr2: Optional[np.ndarray] = None,
         corr2_scale: float = 1.60,
         eps: float = 0.05,
+        quartets: Optional[list] = None,
         pivotal: bool = True,
         single: bool = True,
         remove_fourmembered: bool = True,
@@ -459,56 +472,71 @@ class CFB_ENM(Calculator):
         _dev = self.device
         _tdt = self.torch_dtype
 
-        numbers = images[0].arrays['numbers']
-        r_cov_np = covalent_radii[numbers][:,None]+covalent_radii[numbers]
-        self.r_cov_t = torch.as_tensor(r_cov_np,dtype=_tdt,device=_dev)
[…]
