+++ src/dmf/torch/dmf.py
@@ -419,6 +419,11 @@ class VariationalPathOpt(ABC, cyipopt.Problem):
         d2basis = [b.derivative(nu=2) for b in basis]
         self._basis = [basis,d1basis,d2basis]
 
+        # Callback/evaluation caches
+        self._cached_callback_x = None
+        self._positions_ready = False
+        self._state_eval_ready = False
+
         #t-sequences
         if t_eval is None:
             self.set_t_eval(np.linspace(0.0,1.0,2*self.nsegs+1))
@@ -556,6 +561,37 @@ class VariationalPathOpt(ABC, cyipopt.Problem):
             for b in self._basis[nu]]
             for nu in range(3)])
 
+    def _invalidate_eval_cache(self):
+        self._state_eval_ready = False
+
+    def _invalidate_callback_x_cache(self):
+        self._cached_callback_x = None
+
+    def _ensure_current_state(self):
+        if not self._positions_ready:
+            self.set_positions()
+        if not self._state_eval_ready:
+            self.get_forces()
+            self._state_eval_ready = True
+
+    def _set_x_if_changed(self, x: np.ndarray):
+        x_arr = np.asarray(x)
+        if (
+            self._cached_callback_x is not None
+            and self._cached_callback_x.shape == x_arr.shape
+            and np.array_equal(self._cached_callback_x, x_arr)
+        ):
+            return
+
+        self.set_x(x_arr)
+        if (
+            self._cached_callback_x is None
+            or self._cached_callback_x.shape != x_arr.shape
+        ):
+            self._cached_callback_x = x_arr.copy()
+        else:
+            self._cached_callback_x[...] = x_arr
+
     def set_t_eval(self,t_eval):
         """
         Set the energy evaluation points ``t_eval``.
@@ -577,6 +613,9 @@ class VariationalPathOpt(ABC, cyipopt.Problem):
         self._t_eval_t = torch.as_tensor(
             self.t_eval, dtype=self.torch_dtype, device=self.device
         )
+        self._positions_ready = False
+        self._invalidate_eval_cache()
+        self._invalidate_callback_x_cache()
 
     def set_w_eval(self, w_eval: Optional[np.ndarray] = None):
         """
@@ -603,6 +642,7 @@ class VariationalPathOpt(ABC, cyipopt.Problem):
             w[-1] = 0.5*(self.t_eval[-1]-self.t_eval[-2])
             w[1:-1] = 0.5*(self.t_eval[2:]-self.t_eval[:-2])
             self.w_eval = w
+        self._invalidate_callback_x_cache()
 
     def _get_coefs_from_ref_images(self, ref_images) -> np.ndarray:
         ref_images_copy = [image.copy() for image in ref_images]
@@ -736,6 +776,9 @@ class VariationalPathOpt(ABC, cyipopt.Problem):
         R=self._get_rot_mats()
         self.coefs[-1]=self._coefs0[-1]@R[0]@R[1]@R[2]
         self.coefs_t[-1].copy_(torch.as_tensor(self.coefs[-1], dtype=self.torch_dtype, device=self.device))
+        self._positions_ready = False
+        self._invalidate_eval_cache()
+        self._invalidate_callback_x_cache()
 
     def _get_positions_torch(self, P_t: torch.Tensor, nu=0) -> torch.Tensor:
         return torch.tensordot(P_t[nu].T.contiguous(),self.coefs_t,1)
@@ -780,6 +823,8 @@ class VariationalPathOpt(ABC, cyipopt.Problem):
         pos = self.get_positions()
         for i in range(self.t_eval.size):
             self.images[i].set_positions(pos[i])
+        self._positions_ready = True
+        self._invalidate_eval_cache()
 
     def _get_consts_trans(self) -> np.ndarray:
[…]
