+++ src/dmf/torch/fbenm.py
@@ -549,6 +549,30 @@ class CFB_ENM(Calculator):
             self.quartets_t = torch.as_tensor(self.quartets,dtype=torch.long,device=_dev)
         else:
             self.quartets_t = torch.zeros((0,4),dtype=torch.long,device=_dev)
+        if self.quartets_t.numel() > 0:
+            self._q_i_t = self.quartets_t[:, 0]
+            self._q_j_t = self.quartets_t[:, 1]
+            self._q_k_t = self.quartets_t[:, 2]
+            self._q_l_t = self.quartets_t[:, 3]
+            self._d00_ij_t = self.d_corr0_t[self._q_i_t, self._q_j_t]
+            self._d00_kl_t = self.d_corr0_t[self._q_k_t, self._q_l_t]
+            self._d10_ij_t = self.d_corr1_t[self._q_i_t, self._q_j_t] - self._d00_ij_t
+            self._d10_kl_t = self.d_corr1_t[self._q_k_t, self._q_l_t] - self._d00_kl_t
+            self._d20_ij_t = self.d_corr2_t[self._q_i_t, self._q_j_t] - self._d00_ij_t
+            self._d20_kl_t = self.d_corr2_t[self._q_k_t, self._q_l_t] - self._d00_kl_t
+            self._dnm_t = self._d20_ij_t * self._d20_kl_t - self._d10_ij_t * self._d10_kl_t
+        else:
+            self._q_i_t = torch.zeros((0,), dtype=torch.long, device=_dev)
+            self._q_j_t = torch.zeros((0,), dtype=torch.long, device=_dev)
+            self._q_k_t = torch.zeros((0,), dtype=torch.long, device=_dev)
+            self._q_l_t = torch.zeros((0,), dtype=torch.long, device=_dev)
+            self._d00_ij_t = torch.zeros((0,), dtype=_tdt, device=_dev)
+            self._d00_kl_t = torch.zeros((0,), dtype=_tdt, device=_dev)
+            self._d10_ij_t = torch.zeros((0,), dtype=_tdt, device=_dev)
+            self._d10_kl_t = torch.zeros((0,), dtype=_tdt, device=_dev)
+            self._d20_ij_t = torch.zeros((0,), dtype=_tdt, device=_dev)
+            self._d20_kl_t = torch.zeros((0,), dtype=_tdt, device=_dev)
+            self._dnm_t = torch.zeros((0,), dtype=_tdt, device=_dev)
 
     def copy(self, images=None):
         return type(self)(
@@ -640,19 +664,22 @@ class CFB_ENM(Calculator):
             return
 
         pos = torch.as_tensor(r, dtype=self.torch_dtype, device=self.device)
-        q = self.quartets_t
-        i = q[:,0]; j = q[:,1]; k = q[:,2]; l = q[:,3]
+        i = self._q_i_t
+        j = self._q_j_t
+        k = self._q_k_t
+        l = self._q_l_t
 
         diff_ij = pos[i]-pos[j]
         diff_kl = pos[k]-pos[l]
         d_ij = torch.linalg.norm(diff_ij,dim=1)
         d_kl = torch.linalg.norm(diff_kl,dim=1)
 
-        d00_ij = self.d_corr0_t[i,j]; d00_kl = self.d_corr0_t[k,l]
-        d10_ij = self.d_corr1_t[i,j]-d00_ij
-        d10_kl = self.d_corr1_t[k,l]-d00_kl
[…]
