### 2026-02-12 809b1f3
refactored dmf.torch
