+++ src/dmf/torch/fbenm.py
@@ -1,4 +1,4 @@
-from typing import Optional, Union
+from typing import Optional, Tuple, Union
 
 import numpy as np
 
@@ -7,55 +7,7 @@ from ase.data import covalent_radii
 from ase.data.vdw_alvarez import vdw_radii
 
 import torch
-import warnings
-
-def _resolve_torch_device(device_spec):
-    """
-    Resolve a torch device from a user spec.
-    """
-    if device_spec is None:
-        d = torch.device("cuda" if torch.cuda.is_available() else "cpu")
-    else:
-        try:
-            d = torch.device(device_spec)
-        except (TypeError, ValueError):
-            warnings.warn(f"[dmf.torch] Invalid device spec '{device_spec}', falling back to CPU.")
-            d = torch.device("cpu")
-    if d.type == "cuda" and not torch.cuda.is_available():
-        warnings.warn("[dmf.torch] CUDA requested but not available; falling back to CPU.")
-        d = torch.device("cpu")
-    return d
-
-
-def _resolve_torch_dtype(dtype_spec):
-    """
-    Resolve a torch dtype from a user spec (float32/float64).
-    """
-    if dtype_spec is None:
-        return torch.float64
-    if isinstance(dtype_spec, torch.dtype):
-        if dtype_spec in (torch.float32, torch.float64):
-            return dtype_spec
-        warnings.warn(f"[dmf.torch] Unsupported dtype '{dtype_spec}', falling back to float64.")
-        return torch.float64
-    if isinstance(dtype_spec, str):
-        key = dtype_spec.strip().lower()
-        if key in ("float32", "fp32", "single", "float"):
-            return torch.float32
-        if key in ("float64", "fp64", "double"):
-            return torch.float64
-        warnings.warn(f"[dmf.torch] Invalid dtype spec '{dtype_spec}', falling back to float64.")
-        return torch.float64
-    try:
-        np_dtype = np.dtype(dtype_spec)
-        if np_dtype.kind == "f" and np_dtype.itemsize == 4:
-            return torch.float32
-        if np_dtype.kind == "f" and np_dtype.itemsize == 8:
-            return torch.float64
-    except Exception:
-        pass
-    warnings.warn(f"[dmf.torch] Invalid dtype spec '{dtype_spec}', falling back to float64.")
-    return torch.float64
+from ._torch_config import _resolve_torch_device, _resolve_torch_dtype
 
 
 @torch.no_grad()
@@ -163,6 +115,22 @@ class FB_ENM(Calculator):
             dtype=self.torch_dtype,
         )
 
+    def _ensure_cached_constants(self):
+        specs = (
+            ("_d_min_t", self.d_min),
+            ("_d_max_t", self.d_max),
+            ("_delta_min_t", self.delta_min),
+            ("_delta_max_t", self.delta_max),
+        )
+        for attr, arr in specs:
+            t = getattr(self, attr)
+            if (t is None) or (t.device != self.device) or (t.dtype != self.torch_dtype):
+                setattr(
+                    self,
+                    attr,
+                    torch.as_tensor(arr, dtype=self.torch_dtype, device=self.device),
+                )
+
     @torch.no_grad()
     def calculate(self, atoms, properties, system_changes):
         super().calculate(atoms, properties, system_changes)
@@ -173,14 +141,7 @@ class FB_ENM(Calculator):
         d = torch.cdist(pos_t, pos_t)
 
[…]
