### 2026-02-12 8a735ba
modify t_eval
