+++ src/dmf/torch/fbenm.py
@@ -42,11 +42,18 @@ class FB_ENM(Calculator):
         Scale factor used when delta arrays are autogenerated.
     return_energy_mats : bool
         If True, include ``emat_rep`` and ``emat_att`` in results.
+        Unlike the NumPy implementation, these matrices are omitted by default.
     device : str or torch.device, optional
         Torch device used only for the distance-matrix evaluation.
     dtype : str or torch.dtype, optional
         Torch floating-point dtype for internal calculations
         (``float32`` or ``float64``). Default: ``float64``.
+
+    Notes
+    -----
+    - ``results['energy']`` and ``results['forces']`` are returned as CPU scalars/NumPy arrays.
+    - When ``return_energy_mats=True``, ``results['emat_rep']`` and
+      ``results['emat_att']`` are returned as Torch tensors on ``device``.
 
     """
 
