+++ src/dmf/torch/fbenm.py
@@ -76,6 +76,9 @@ class FB_ENM(Calculator):
         self.torch_dtype = _resolve_torch_dtype(dtype)
         self.np_dtype = np.float64
         self._return_energy_mats = bool(return_energy_mats)
+        # CUDA memory is typically the bottleneck for large systems.
+        # Keep constants cached on device only on CPU backend.
+        self._cache_constants_on_device = (self.device.type != "cuda")
         self._d_min_t = None
         self._d_max_t = None
         self._delta_min_t = None
@@ -123,6 +126,8 @@ class FB_ENM(Calculator):
         )
 
     def _ensure_cached_constants(self):
+        if not self._cache_constants_on_device:
+            return
         specs = (
             ("_d_min_t", self.d_min),
             ("_d_max_t", self.d_max),
@@ -138,6 +143,25 @@ class FB_ENM(Calculator):
                     torch.as_tensor(arr, dtype=self.torch_dtype, device=self.device),
                 )
 
+    def _get_constants_t(self):
+        if self._cache_constants_on_device:
+            self._ensure_cached_constants()
+            return self._d_min_t, self._d_max_t, self._delta_min_t, self._delta_max_t
+        return (
+            torch.as_tensor(self.d_min, dtype=self.torch_dtype, device=self.device),
+            torch.as_tensor(self.d_max, dtype=self.torch_dtype, device=self.device),
+            torch.as_tensor(self.delta_min, dtype=self.torch_dtype, device=self.device),
+            torch.as_tensor(self.delta_max, dtype=self.torch_dtype, device=self.device),
+        )
+
+    def release_device_cache(self, empty_cache: bool = False):
+        self._d_min_t = None
+        self._d_max_t = None
+        self._delta_min_t = None
+        self._delta_max_t = None
+        if empty_cache and self.device.type == "cuda" and torch.cuda.is_available():
+            torch.cuda.empty_cache()
+
     @torch.no_grad()
     def calculate(self, atoms, properties, system_changes):
         super().calculate(atoms, properties, system_changes)
@@ -147,15 +171,14 @@ class FB_ENM(Calculator):
         pos_t = torch.as_tensor(pos, dtype=self.torch_dtype, device=self.device)
         d = torch.cdist(pos_t, pos_t)
 
-        # Cache torch copies of constants on the target device/dtype
-        self._ensure_cached_constants()
+        d_min_t, d_max_t, delta_min_t, delta_max_t = self._get_constants_t()
 
         # 2) Flat-bottom penalties (Torch on device)
-        d_rep = torch.clamp(d - self._d_min_t, max=0.0)
-        d_att = torch.clamp(d - self._d_max_t, min=0.0)
+        d_rep = torch.clamp(d - d_min_t, max=0.0)
+        d_att = torch.clamp(d - d_max_t, min=0.0)
 
-        inv_dmin2 = 1.0 / (self._delta_min_t * self._delta_min_t)
-        inv_dmax2 = 1.0 / (self._delta_max_t * self._delta_max_t)
+        inv_dmin2 = 1.0 / (delta_min_t * delta_min_t)
+        inv_dmax2 = 1.0 / (delta_max_t * delta_max_t)
 
         if self._return_energy_mats:
             e_rep = (d_rep * d_rep) * inv_dmin2
@@ -188,6 +211,9 @@ class FB_ENM(Calculator):
[…]
