### 2026-02-13 1ba8edb
modify device selection
