### 2026-03-20 #9806 @abhishektiwari
👋 @naik-aakash @ml-evs @jacksund, would any of you be willing to review this submission for Journal of Open Source Software (JOSS)? The submission under consideration is "PyDMF: A Python package for variational double-ended reaction-path and transition-state optimization".

Thanks
Abhi
