### 2026-03-31 #9806 @abhishektiwari
@staradutt @berquist @RiesBen  would any of you be willing to review this submission for Journal of Open Source Software (JOSS)? The submission under consideration is "PyDMF: A Python package for variational double-ended reaction-path and transition-state optimization".

Thanks
Abhi
