### 2026-04-06 #10379 @editorialbot
[REVIEW]: PyDMF: A Python package for variational double-ended reaction-path and transition-state optimization
**Submitting author:** @shin1koda (<a href="http://orcid.org/0000-0003-0993-3678">Shin-ichi Koda</a>)
**Repository:** https://github.com/shin1koda/dmf
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0.0
