### 2026-04-06 #10379 @berquist
- acceptable; MIT

  - ok
  - ok
  - ok; missing release process/tags

  - used pixi instead of conda
  - `sample/sample.py` and tests run without issue

  - good

  - good
  - good
  - good
  - ok
  - good
