### 2026-04-14 shin1koda/dmf#5 @naik-aakash
Suggestion: update `pyproject.toml`
This issue is part of JOSS review https://github.com/openjournals/joss-reviews/issues/10379

I had a quick look at the current ASE package Python version requirement.

It seems it needs at least version >=3.10, https://gitlab.com/ase/ase/-/blob/master/pyproject.toml?ref_type=heads

Thus, I would suggest it would be good to also update the same repositories `pyproject.toml` file here https://github.com/shin1koda/dmf/blob/5296a254d11331b43146bed7b77b2f655f8d4584/pyproject.toml#L11

Also, it would be nice to explicitly specify the supported Python versions in `pyproject.toml`, as ASE does, for example.

## Update optional dependencies

Add `mpi4py ` under optional dependencies; this is missing
