### 2026-04-14 shin1koda/dmf#6 @naik-aakash
Suggestion : Include a CHANGELOG
This issue is part of JOSS review https://github.com/openjournals/joss-reviews/issues/10379

It would be beneficial in the long run to include a `CHANGELOG` file documenting changes with each release. Currently, this seems to be entirely missing
