### 2026-04-14 shin1koda/dmf#7 @naik-aakash
Suggestion: Update testing workflow
Issue is part of JOSS review https://github.com/openjournals/joss-reviews/issues/10379

1. I suggest testing at least Python versions 3.11 and 3.12 in the testing workflow as well. Currently, automated tests are only run for 3.10

2. Also, it would be nice if more `tests` could be added. Currently, coverage stands 71%.

Command used `pytest tests --cov=src/dmf --cov-report term-missing`

```bash
Name                     Stmts   Miss  Cover   Missing
------------------------------------------------------
src/dmf/__init__.py          8      2    75%   43-44
src/dmf/dmf.py             588    125    79%   18, 260-266, 275, 300-301, 304-305, 308-310, 316, 346, 355, 361, 367, 380, 442-443, 577, 744, 816-829, 833, 844-882, 886-887, 922, 950, 988, 1166, 1173-1182, 1185-1196, 1235-1241, 1781-1823
src/dmf/fbenm.py           259     81    69%   200, 211, 214, 221, 225, 395, 398-433, 452-476, 485-490, 493-495, 498-499, 502-503, 506-509, 512-515, 536-537, 540, 550-553, 588-594
src/dmf/interpolate.py      93     53    43%   83, 94-102, 120, 132, 144, 219-297
src/dmf/mpi_backend.py      16     15     6%   11-36
------------------------------------------------------
TOTAL                      964    276    71%
```
