### 2026-04-19 shin1koda/dmf#8 @naik-aakash
Suggestion on  README
Suggestion is part of JOSS review https://github.com/openjournals/joss-reviews/issues/10379

It would be beneficial if the authors also include the link to the work they refer in the paper that uses the methods from this package in MLIPs

https://pubs.acs.org/doi/10.1021/acs.jcim.5c02398
