### 2026-04-19 shin1koda/dmf#9 @naik-aakash
Update documentation
Suggestion is part of JOSS review https://github.com/openjournals/joss-reviews/issues/10379

It would be essential to include the README content on the API documentation page.

Also, please include an example of using the parallel option in DirectMaxFlux. Setting `parallel` arg to `mpi` does not seem to work. It fails with a runtime error

`RuntimeError: parallel='mpi' requires calc_factory`

## Reasons

1. Any non-expert potential user who lands on the package website will not currently see any installation/usage instructions.
2. There is no link pointing the user to the repository's source to figure this out.
