### 2026-04-19 shin1koda/dmf#10 @naik-aakash
release tags not available on the repository
Suggestion is part of JOSS review https://github.com/openjournals/joss-reviews/issues/10379

Ideally, a code repository has tags associated with each new release of the Python package. Currently, it is missing. Not sure how the PyPI releases are currently being handled.

Adding a workflow to the repository to handle PyPI releases would keep things in sync; the same goes for documentation deployment on GitHub Pages.
