+++ src/dmf/interpolate.py
@@ -71,151 +71,6 @@ def interpolate_fbenm(
     - For details of FB-ENM and CFB-ENM, see the corresponding papers.
     """
 
-
-    mxflx = DirectMaxFlux(ref_images,
-                          nmove=nmove,
-                          update_teval=False,
-                          **dmf_options)
-
-    if fbenm_only_endpoints:
-        fbenm_images = [ref_images[0].copy(),ref_images[-1].copy()]
-    else:
-        fbenm_images = [image.copy() for image in ref_images]
-
-    if copy_calc0:
-        calc_f = FB_ENM_Bonds(fbenm_images, **fbenm_options)
-        if correlated:
-            calc_c = CFB_ENM(fbenm_images,**cfbenm_options)
-        for image in mxflx.images:
-            if correlated:
-                image.calc = SumCalculator([calc_f.copy(),
-                                            calc_c.copy(fbenm_images)])
-            else:
-                image.calc = calc_f.copy()
-    else:
-        for i,image in enumerate(mxflx.images):
-            if correlated:
-                image.calc = SumCalculator([
-                                 FB_ENM_Bonds(fbenm_images, **fbenm_options),
-                                 CFB_ENM(fbenm_images,**cfbenm_options)])
-            else:
-                image.calc = FB_ENM_Bonds(fbenm_images, **fbenm_options)
-
-    options ={
-        'tol': 0.1,
-        'dual_inf_tol': 0.01,
-        'constr_viol_tol': 0.01,
-        'compl_inf_tol': 0.01,
-        'nlp_scaling_method':'user-scaling',
-        'obj_scaling_factor':0.1,
-        'limited_memory_initialization':'constant',
-        'limited_memory_init_val':2.5,
-        'accept_every_trial_step':'yes',
-        'output_file':output_file,
-        'max_iter':200,
-        }
-    mxflx.add_ipopt_options(options)
-
-    if ipopt_options:
-        mxflx.add_ipopt_options(ipopt_options)
-
-    if sequential:
-        b_scale = 3.0
-        w_eval0 = mxflx.w_eval.copy()
-        for i in range((nmove+1)//2):
-            mxflx.get_forces()
-            ens = mxflx.energies.copy()
-            w_eval = w_eval0.copy()
-            ens[i+2:nmove-i]=0.0
-            w_eval[i+2:nmove-i]=0.0
-            if np.amax(ens)>0.0:
-                mxflx.beta=b_scale/np.amax(ens)
-            else:
-                mxflx.beta=1.0
-            mxflx.set_w_eval(w_eval)
-
-            mxflx.solve(tol=0.1)
-
-    b_scale = 5.0
-    for _ in range(5):
-        mxflx.get_forces()
-        ens = mxflx.energies.copy()
-        if np.amax(ens)>0.0:
-            mxflx.beta = b_scale/np.amax(ens)
-        else:
-            mxflx.beta = 1.0
-
-        mxflx.solve(tol=0.1)
-
-    return mxflx
-
-def interpolate_fbenm_new(
-        ref_images,nmove=10,
-        output_file='fbenm_ipopt.out',
-        correlated=True,
-        sequential=True,
-        fbenm_only_endpoints=True,
-        copy_calc0=True,
-        fbenm_options={},
-        cfbenm_options={},
-        dmf_options={},
-        ipopt_options={},
-        ):
-    """
-    Generate a plausible initial reaction path using FB-ENM or
-    FB-ENM + CFB-ENM in combination with the direct MaxFlux method.
-
-    This routine constructs a DirectMaxFlux object from the given
-    reference images (typically reactant and product) and assigns an
-    FB-ENM_Bonds calculator, optionally combined with CFB-ENM,
-    to each intermediate image. The DMF solver is then executed with
[…]
