+++ tests/test_dmf.py
@@ -29,3 +29,35 @@ def test_interpolate_fbenm_basic():
         atol=1e-8
     )
 
+def test_maxflux():
+
+    import numpy as np
+    from ase.io import write, read
+    from ase.calculators.emt import EMT
+    from dmf import DirectMaxFlux, interpolate_fbenm
+
+    # read react.xyz and prod.xyz
+    ref_images = [read('sample/react.xyz'), read('sample/prod.xyz')]
+
+    # generate initial path by FB-ENM
+    mxflx_fbenm = interpolate_fbenm(ref_images,correlated=True)
+
+    ## write initial path and its coefficients
+    #write('sample_ini.traj',mxflx_fbenm.images)
+    coefs = mxflx_fbenm.coefs.copy()
+    #np.save('sample_ini_coefs',coefs)
+
+    # set up a variational problem of the direct MaxFlux method
+    mxflx = DirectMaxFlux(ref_images,coefs=coefs,nmove=3,update_teval=True)
+
+    # set up calculators
+    for image in mxflx.images:
+        image.calc = EMT()
+
+    # solve the variational problem
+    mxflx.add_ipopt_options({'output_file':'tests/sample_ipopt.out'})
+    mxflx.solve(tol='middle')
+
+    ## write final path and history of x(tmax)
+    #write('sample_fin.traj',mxflx.images)
+    #write('sample_tmax.traj',mxflx.history.images_tmax)
