+++ CHANGELOG.md
@@ -1,0 +1,14 @@
+# Changelog
+
+All notable changes to this project will be documented in this file.
+
+## [1.1.1] - 2026-04-21
+
+### Added
+- Added `CHANGELOG.md` (this file).
+- Added `mpi4py` under optional dependencies.
+- Added `.gitignore` files to `tests/` and `sample/` directories.
+
+### Changed
+- Updated `requires-python` to `>=3.10`.
+
