+++ .github/workflows/tests.yml
@@ -7,6 +7,9 @@ on:
 jobs:
   test:
     runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: [ "3.10", "3.11", "3.12" ]
     defaults:
       run:
         shell: bash -el {0}
@@ -19,7 +22,7 @@ jobs:
           micromamba-version: 'latest'
           environment-name: pydmf
           create-args: >-
-            python=3.10
+            python=${{ matrix.python-version }}
           cache-environment: true
 
       - name: Install dependencies
