+++ tests/test_dmf.py
@@ -41,11 +41,7 @@ def test_maxflux():
 
     # generate initial path by FB-ENM
     mxflx_fbenm = interpolate_fbenm(ref_images,correlated=True)
-
-    ## write initial path and its coefficients
-    #write('sample_ini.traj',mxflx_fbenm.images)
     coefs = mxflx_fbenm.coefs.copy()
-    #np.save('sample_ini_coefs',coefs)
 
     # set up a variational problem of the direct MaxFlux method
     mxflx = DirectMaxFlux(ref_images,coefs=coefs,nmove=3,update_teval=True)
@@ -57,7 +53,3 @@ def test_maxflux():
     # solve the variational problem
     mxflx.add_ipopt_options({'output_file':'tests/sample_ipopt.out'})
     mxflx.solve(tol='middle')
-
-    ## write final path and history of x(tmax)
-    #write('sample_fin.traj',mxflx.images)
-    #write('sample_tmax.traj',mxflx.history.images_tmax)
