### 2026-04-23 e0369b2
add Related works (issue #8)
