### 2026-05-21 12bcf47
add a github workflow
