### 2026-05-21 fe8336f
Add PyTorch Backend (`dmf.torch`) with Optional GPU Acceleration (#3)
* add torch_dmf

* add torch_dmf

* update default

* add constraint

* add constraint

* update

* add guard for t_eval update

* update torch_dmf

* update torch_dmf

* remove constraint

* remove np.float32

* refactoring

* refactoring

* refactored dmf.torch

* modify t_eval

* update dmf.torch

* modify COM

* modify device selection

---------

Co-authored-by: Dokkoisyo <chemistry.waseda@gmail.com>
