+++ src/dmf/torch/interpolate.py
@@ -1,0 +1,348 @@
+import numpy as np
+from ase.calculators.mixing import SumCalculator
+from .dmf import DirectMaxFlux
+from .fbenm import FB_ENM_Bonds, CFB_ENM
+
+
+def _extract_common_kwargs(fbenm_options, cfbenm_options, dmf_options):
+    dmf_options = dict(dmf_options) if dmf_options is not None else {}
+    fbenm_options = dict(fbenm_options) if fbenm_options is not None else {}
+    cfbenm_options = dict(cfbenm_options) if cfbenm_options is not None else {}
+
+    device = dmf_options.pop('device', None)
+    device = fbenm_options.pop('device', device)
+    device = cfbenm_options.pop('device', device)
+
+    dtype = dmf_options.pop('dtype', None)
+    dtype = fbenm_options.pop('dtype', dtype)
+    dtype = cfbenm_options.pop('dtype', dtype)
+
+    common_kwargs = {}
+    if device is not None:
+        common_kwargs['device'] = device
+    if dtype is not None:
+        common_kwargs['dtype'] = dtype
+
+    return fbenm_options, cfbenm_options, dmf_options, common_kwargs
+
+
+def interpolate_fbenm(
+        ref_images,nmove=10,
+        output_file='fbenm_ipopt.out',
+        correlated=True,
+        sequential=True,
+        fbenm_only_endpoints=True,
+        copy_calc0=True,
+        fbenm_options=None,
+        cfbenm_options=None,
+        dmf_options=None,
+        ipopt_options=None,
+        device=None,
+        ):
+    """
+    Generate a plausible initial reaction path using FB-ENM or
+    FB-ENM + CFB-ENM in combination with the direct MaxFlux method.
+
+    This routine constructs a DirectMaxFlux object from the given
+    reference images (typically reactant and product) and assigns an
+    FB-ENM_Bonds calculator, optionally combined with CFB-ENM,
+    to each intermediate image. The DMF solver is then executed with
+    a β-update scheme (see FB-ENM's paper) to obtain a plausible path.
+
+    Parameters
+    ----------
+    ref_images : list of ase.Atoms
+        Reference structures defining the initial piecewise linear path
+        for the (C)FB-ENM optimization.
+    nmove : int, optional
+        Number of movable images. Default: 10.
+    output_file : str, optional
+        File name for IPOPT output. Default: 'fbenm_ipopt.out'.
+    correlated : bool, optional
+        If True, use FB-ENM + CFB-ENM (correlated ENM).
+        If False, use FB-ENM only. Default: True.
+    sequential : bool, optional
+        Whether to apply a sequential MaxFlux optimization scheme
+        that gradually activates interior points. Default: True.
+    fbenm_only_endpoints : bool, optional
+        If True, construct FB-ENM from only the first and last image.
+        If False, use all ref_images for FB-ENM construction.
+        Default: True.
+    copy_calc0 : bool, optional
+        If True, the calculator of images[0] is copied to the other images.
+        If False, independently initialize each calculator.
+        Default: True.
+    fbenm_options : dict, optional
+        Keyword arguments forwarded to `FB_ENM_Bonds`.
+    cfbenm_options : dict, optional
+        Keyword arguments forwarded to `CFB_ENM`.
+    dmf_options : dict, optional
+        Keyword arguments forwarded to `DirectMaxFlux`.
+    ipopt_options : dict, optional
+        Keyword arguments forwarded to `DirectMaxFlux.add_ipopt_options()`.
+    device : str or torch.device, optional
+        Torch device for internal tensors (e.g. ``"cuda"``).
[…]
