+++ src/dmf/torch/dmf.py
@@ -1,0 +1,2125 @@
+import threading
+from abc import ABC, abstractmethod
+from collections import defaultdict
+from typing import Optional
+
+import numpy as np
+from numpy.polynomial import polynomial as P
+from scipy.interpolate import BSpline
+from scipy.spatial.transform import Rotation
+import cyipopt
+
+from functools import cached_property
+
+import torch
+from ._torch_config import _resolve_torch_device, _resolve_torch_dtype
+
+from ..mpi_backend import _worker_loop
+try:
+    import os
+    os.environ['UCX_LOG_LEVEL'] = 'error'
+    from mpi4py import MPI
+    HAS_MPI4PY = True
+except ImportError:
+    HAS_MPI4PY = False
+
+
+@torch.no_grad()
+def _interp1d_torch(xq: torch.Tensor, xp: torch.Tensor, fp: torch.Tensor) -> torch.Tensor:
+    """
+    Linear interpolate fp(xp) at xq using pure PyTorch.
+
+    xp: (M,), fp: (M, *), xq: (K,)
+    Returns: (K, *) broadcast along the trailing dims of fp.
+    """
+    idx = torch.searchsorted(xp,xq,right=False).clamp(1,xp.numel()-1)
+    x0, x1 = xp[idx-1], xp[idx]
+    lam = (xq-x0)/(x1-x0)
+    while lam.dim() < fp.dim():
+        lam = lam.unsqueeze(-1)
+    f0, f1 = fp[idx-1], fp[idx]
+    return f0 + (f1-f0)*lam
+
+
+def _release_calc_device_cache(calc, *, empty_cache: bool = False):
+    if calc is None:
+        return
+
+    release = getattr(calc, "release_device_cache", None)
+    if callable(release):
+        try:
+            release(empty_cache=empty_cache)
+        except TypeError:
+            release()
+
+    mixer = getattr(calc, "mixer", None)
+    if mixer is not None:
+        sub_calcs = getattr(mixer, "calcs", None)
+        if sub_calcs is not None:
+            for sub in sub_calcs:
+                _release_calc_device_cache(sub, empty_cache=False)
+
+
+class HistoryBase():
+    """
+    Container storing the optimization history of the VariationalPathOpt.
+
+    This object collects various physical and numerical quantities evaluated
+    along the reaction path during the optimization.  At each IPOPT iteration,
+    the ``VariationalPathOpt.intermediate`` method appends the current values
+    of these quantities to the corresponding lists below.
+
+    Attributes
+    ----------
+    forces : list of ndarray
+        History of ``VariationalPathOpt.forces``.
+    energies : list of ndarray
+        History of ``VariationalPathOpt.energies``.
+    coefs : list of ndarray
+        History of ``VariationalPathOpt.coefs``.
+    angs : list of ndarray
+        History of ``VariationalPathOpt.angs``.
+    tmax : list of float
+        History of the location ``t_max`` corresponding to the maximum
+        interpolated energy along the path. See Ref. 1 for details.
+    images_tmax : list of ase.Atoms
+        History of the atomic structure at ``t = t_max``, providing an
+        approximate transition-state geometry at each iteration.
+    duals : list of float
+        History of the scaled dual infeasibility (IPOPT diagnostic).
+
+    """
+
+    def __init__(self):
+        self.forces = []
+        self.energies = []
+        self.coefs = []
+        self.angs = []
+        self.tmax = []
+        self.images_tmax = []
+        self.duals = []
+
+
+class VariationalPathOpt(ABC, cyipopt.Problem):
+    r"""
+    Abstract base class for variational reaction–path optimization.
+
+    This class formulates a general functional
+
+    .. math::
+
[…]
