+++ README.md
@@ -91,6 +91,32 @@ Running the script produces several output files, including:
  - `sample_tmax.traj`: History of the estimated highest-energy point during optimization
 
 
+## dmf.torch (GPU acceleration)
+
+`dmf.torch` is a PyTorch-accelerated backend that mirrors the `dmf` API while offloading internal tensor operations to PyTorch (CUDA when available).
+
+If you want to use `dmf.torch`, install a CUDA-matched build of PyTorch before installing PyDMF.  
+Example for CUDA 12.8:
+
+```bash
+pip install torch --index-url https://download.pytorch.org/whl/cu128
+pip install "pydmf[torch]"
+```
+
+To use the PyTorch backend, import from `dmf.torch`. The rest of the API is unchanged:   
+
+```python
+from dmf.torch import DirectMaxFlux, interpolate_fbenm
+```
+
+When using `dmf.torch`, it automatically utilizes the available CUDA device. However, you can optionally specify the CUDA device in entry points such as `DirectMaxFlux` and `interpolate_fbenm`:
+
+```python
+mxflx = DirectMaxFlux(ref_images, coefs=coefs, nmove=3, device="cuda")
+mxflx_fbenm = interpolate_fbenm(ref_images, correlated=True, device="cuda")
+```
+
+
 ## Documentation
 
 For more details, please refer to the [API documentation](https://shin1koda.github.io/dmf/).
@@ -146,4 +172,3 @@ We will do our best to provide guidance based on availability.
 ## License
 
 This project is distributed under the MIT License.
-

[…]
