### 2026-05-21 0b54984
v1.2.0 GPU acceleration
