### 2026-05-21 bdfb5ae
v1.2.0 GPU acceleration
