### 2026-05-26 cc05f04
remove MPI-based parallelization
