+++ src/dmf/dmf.py
@@ -1,6 +1,5 @@
 import threading
 from abc import ABC, abstractmethod
-from collections import defaultdict
 
 import numpy as np
 from numpy.polynomial import polynomial as P
@@ -9,15 +8,6 @@ from scipy.spatial.transform import Rotation
 import cyipopt
 
 from functools import cached_property
-
-from .mpi_backend import _worker_loop
-try:
-    import os
-    os.environ['UCX_LOG_LEVEL'] = 'error'
-    from mpi4py import MPI
-    HAS_MPI4PY = True
-except ImportError:
-    HAS_MPI4PY = False
 
 class HistoryBase():
     """
@@ -247,7 +237,7 @@ class VariationalPathOpt(ABC, cyipopt.Problem):
         remove_rotation_and_translation=True,
         mass_weighted=False,
         calc_factory=None,
-        parallel=False,world=None,
+        parallel=False,
         t_eval=None,w_eval=None,
         n_vel=None,n_trans=None,n_rot=None,
         eps_vel=0.01,eps_rot=0.01,
@@ -255,20 +245,6 @@ class VariationalPathOpt(ABC, cyipopt.Problem):
 
         #Prallel calculation
         self.parallel = parallel
-
-        if parallel and (parallel == "mpi" or world is not None):
-            self.parallel = "mpi"
-            if not HAS_MPI4PY:
-                raise RuntimeError("MPI parallel calculation requires mpi4py.")
-
-            self._world = MPI.COMM_WORLD if world is None else world
-            self._rank = self._world.Get_rank()
-            self._size = self._world.Get_size()
-        else:
-            self._world = None
-            self._rank = 0
-            self._size = 1
-
 
         #Initialize images
         if t_eval is None:
@@ -280,35 +256,12 @@ class VariationalPathOpt(ABC, cyipopt.Problem):
         for _ in range(self._nimages):
             self.images.append(ref_images[0].copy())
 
-        r2i = defaultdict(list)
-        for i in range(self._nimages):
-            if i==0:
-                r = 0
-            elif i==self._nimages-1:
-                r = self._size-1
-            else:
-                r = (i-1)%self._size
-
-            r2i[r].append(i)
-
-        self._r2i = r2i
-
         #calc_factory
         self.calc_factory = calc_factory
 
-        if self.parallel == "mpi":
-            if self.calc_factory is None:
-                raise RuntimeError("parallel='mpi' requires calc_factory.")
-
         if self.calc_factory is not None:
-            for i in self._r2i[self._rank]:
-                self.images[i].calc = self.calc_factory(i)
-
-        if self.parallel == "mpi" and self._rank>0:
-            _worker_loop(self._world, self.images)
-            import sys
-            sys.exit(0)
-
+            for i, image in enumerate(self.images):
+                image.calc = self.calc_factory(i)
 
         #Atoms
         self.natoms = len(ref_images[0])
@@ -437,10 +390,6 @@ class VariationalPathOpt(ABC, cyipopt.Problem):
             'accept_every_trial_step':'yes',
             'output_file':'pathopt.out',
             }
-
-        if self.parallel=='mpi':
-            if self._rank > 0:
-                defaults['print_level'] = 0
 
         self.ipopt_options = dict()
         self.add_ipopt_options(defaults)
@@ -811,7 +760,7 @@ class VariationalPathOpt(ABC, cyipopt.Problem):
 
     def _get_forces_by_img_idxs(self,idxs,energies,forces):
 
-        if self.parallel and self._world is None:
+        if self.parallel:
 
             def run(image, energies, forces):
                 forces[:] = image.get_forces()
[…]
