### 2026-05-26 a8d753e
add calc_factory option
