+++ src/dmf/torch/dmf.py
@@ -120,7 +120,7 @@ class VariationalPathOpt(ABC, cyipopt.Problem):
     Additional features include:
     - construction of initial B-spline coefficients from ``ref_images``
     - optional removal of translational and rotational redundancy
-    - parallel energy/force evaluation (threads or MPI)
+    - parallel energy/force evaluation using Python threads
 
 
     Parameters
@@ -156,8 +156,13 @@ class VariationalPathOpt(ABC, cyipopt.Problem):
         If True, the velocity norm \( \vert \dot{x}(t) \vert \) uses mass-weighted
         coordinates. Default: False.
 
+    calc_factory : callable, optional
+        Factory function returning a calculator for image index ``i``.
+        If provided, ``calc_factory(i)`` is assigned to ``images[i].calc``.
+        Default: None.
+
     parallel : bool, optional
-        Evaluate energies and forces in parallel using threads or MPI.
+        Evaluate energies and forces in parallel using Python threads.
         Default: False.
 
     world : MPI communicator, optional
@@ -1568,7 +1573,7 @@ class DirectMaxFlux(VariationalPathOpt):
     Additional features include:
     - construction of initial B-spline coefficients from ``ref_images``
     - optional removal of translational and rotational redundancy
-    - parallel energy/force evaluation (threads or MPI)
+    - parallel energy/force evaluation Python threads
     - optional adaptive refinement of ``t_eval`` near the high-energy region
 
 
@@ -1605,8 +1610,13 @@ class DirectMaxFlux(VariationalPathOpt):
         If True, the velocity norm \( \vert \dot{x}(t) \vert \) uses mass-weighted
         coordinates. Default: False.
 
+    calc_factory : callable, optional
+        Factory function returning a calculator for image index ``i``.
+        If provided, ``calc_factory(i)`` is assigned to ``images[i].calc``.
+        Default: None.
+
     parallel : bool, optional
-        Evaluate energies and forces in parallel using threads or MPI.
+        Evaluate energies and forces in parallel using Python threads.
         Default: False.
 
     world : MPI communicator, optional
