+++ src/dmf/__init__.py
@@ -1,5 +1,5 @@
 """
-PyDMF: Direct MaxFlux and FB-ENM utilities
+PyDMF v1.2.1 : Direct MaxFlux and FB-ENM utilities
 ========================================
 
 This package provides an implementation of the Direct MaxFlux (DMF) method
