### 2026-05-26 808b1ad
add tests for torch
