### 2026-05-27 ebe246e
add tests for torch
