### 2026-05-27 0dada08
add a GPU-support link to docs
