+++ docs/source/conf.py
@@ -47,4 +47,4 @@ autodoc_default_options = {
 # --- Options for HTML output ---------------------------------------------------
 html_theme = 'sphinx_rtd_theme'
 html_static_path = ['_static']
-
+html_logo = "_static/pydmf-logo.png"
