+++ docs/source/index.rst
@@ -10,6 +10,13 @@
 
 
 
+GPU Acceleration
+----------------
+
+GPU acceleration has been supported since v1.2.0. The design of the GPU-accelerated classes and methods is mostly consistent with the standard implementation. For details, see the `README_TORCH.md <https://github.com/shin1koda/dmf/blob/main/README_TORCH.md>`_ file.
+
+
+
 Contents
 --------
 
