+++ _sources/index.rst.txt
@@ -1,0 +1,35 @@
+.. dmf documentation master file
+
+.. automodule:: dmf
+   :no-members:
+   :no-imported-members:
+   :no-special-members:
+   :no-undoc-members:
+   :no-inherited-members:
+   :exclude-members: *
+
+
+
+GPU Acceleration
+----------------
+
+GPU acceleration has been supported since v1.2.0. The design of the GPU-accelerated classes and methods is mostly consistent with the standard implementation. For details, see the `README_TORCH.md <https://github.com/shin1koda/dmf/blob/main/README_TORCH.md>`_ file.
+
+
+
+Contents
+--------
+
+.. toctree::
+   :maxdepth: 2
+
+   dmf
+   fbenm
+   interpolate
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`search`
+
