### 2026-05-27 ad67244
add dmf-g16 as an application of PyDMF
