+++ paper/paper.bib
@@ -1,3 +1,14 @@
+@article{koda2026dmfg16,
+  year     = {2026},
+  title    = {dmf‐g16: A Gaussian Wrapper for Reliable Double‐Ended Transition‐State Searches With Native Input Formats},
+  author   = {Koda, Shin‐ichi and Saito, Shinji},
+  journal  = {Journal of Computational Chemistry},
+  doi      = {10.1002/jcc.70378},
+  pages    = {e70378},
+  number   = {12},
+  volume   = {47}
+}
+
 @article{koda2025correlat,
     author = "Koda, Shin-ichi and Saito, Shinji",
     year = "2025",
