+++ paper/paper.md
@@ -61,8 +61,7 @@ In PyDMF, the implementation focuses on defining the nonlinear optimization prob
 
 The research impact of PyDMF is supported by both quantitative benchmark results and its integration into user-facing computational workflows. The efficiency and robustness of DMF and FB-ENM relative to existing reaction-path optimization approaches have been demonstrated in benchmark studies reported in their original publications [@koda2024locating; @koda2024flatbott; @koda2025correlat]. Using a dataset of 121 representative chemical reactions involving typical elements [@asgeirsson2021nudgedel], DMF reduced the computational cost by approximately 70% compared with conventional NEB calculations. FB-ENM was shown to generate more energetically favorable reaction paths than IDPP and to produce chemically plausible initial paths even for complex reactions where IDPP often fails.
 
-Beyond these benchmarks, PyDMF has been adopted as a core component in higher-level computational workflows. Notably, it serves as the transition-state search engine in ColabReaction, a web-based application for transition-state searches [@karasawa2025colabrea]. This integration demonstrates that PyDMF is not only a methodological implementation but also a practical and reusable software component that supports accessible and reliable transition-state searches.
-
+Beyond these benchmarks, PyDMF has been adopted as a core component in higher-level computational workflows. Notably, it serves as the transition-state search engine in ColabReaction, a web-based application for transition-state searches [@karasawa2025colabrea]. PyDMF is also used as the underlying path-optimization engine in dmf-g16, a Gaussian wrapper that enables reliable double-ended transition-state searches from native Gaussian input files [@koda2026dmfg16]. These integrations demonstrate that PyDMF is not only a methodological implementation but also a practical and reusable software component that supports accessible and reliable transition-state searches.
 
 # AI usage disclosure
 
