### 2026-05-27 012f174
add instruction for parallel calculation
