+++ paper/paper.bib
@@ -1,3 +1,42 @@
+@article{nakano2025high, 
+  year     = {2025}, 
+  title    = {High-Speed Terpene Pathway Analysis via Double-ended Transition State Search Method Combined with {ML} Potentials Uncovers the Intricate Rearrangement Reaction in Spiroalbatene Biosynthesis}, 
+  author   = {Nakano, Moe and Karasawa, Masayuki and Ohmura, Takuto and Terada, Tohru and Sato, Hajime}, 
+  journal  = {ChemRxiv}, 
+  doi      = {10.26434/chemrxiv-2025-md8k6-v2}, 
+}
+
+@article{dhiman2026porcine, 
+  year     = {2026}, 
+  title    = {Porcine Liver Esterase‐Catalyzed Dynamic Kinetic Resolution of a Lactone‐Bridged Biaryl: Impact of Organic Cosolvents on Enantioselectivity}, 
+  author   = {Dhiman, Neha and Hummel, Anke and Kanomata, Kyohei and Sawama, Yoshinari and Taniguchi, Tohru and Gröger, Harald and Akai, Shuji}, 
+  journal  = {European Journal of Organic Chemistry}, 
+  issn     = {1434-193X}, 
+  doi      = {10.1002/ejoc.70473}, 
+  number   = {20}, 
+  volume   = {29}
+}
+
+@article{ohmura2026pdb2, 
+  year     = {2026}, 
+  title    = {pdb2reaction: End-to-End Reaction-Path Elucidation from {PDB} Structures Using Machine-Learning Interatomic Potentials}, 
+  author   = {Ohmura, Takuto and Sato, Hajime and Terada, Tohru}, 
+  journal  = {ChemRxiv}, 
+  doi      = {10.26434/chemrxiv.15003538/v1}, 
+}
+
+@article{kondo2026osim, 
+  year     = {2026}, 
+  title    = {Osimertinib’s Proton-Catalyzed, Pseudoconcerted {EGFR} Inhibition Guides Next-Generation Inhibitor Design}, 
+  author   = {Kondo, Akihiro and Fujimoto, Kazuhiro J. and Koda, Shin-ichi and Saito, Shinji and Yanai, Takeshi}, 
+  journal  = {Journal of Chemical Information and Modeling}, 
+  issn     = {1549-9596}, 
+  doi      = {10.1021/acs.jcim.5c02787}, 
+  pages    = {1203--1213}, 
+  number   = {2}, 
+  volume   = {66}
+}
+
 @article{koda2026dmfg16,
   year     = {2026},
   title    = {dmf‐g16: A Gaussian Wrapper for Reliable Double‐Ended Transition‐State Searches With Native Input Formats},
