+++ paper/paper.md
@@ -59,9 +59,11 @@ In PyDMF, the implementation focuses on defining the nonlinear optimization prob
 
 # Research impact statement
 
-The research impact of PyDMF is supported by both quantitative benchmark results and its integration into user-facing computational workflows. The efficiency and robustness of DMF and FB-ENM relative to existing reaction-path optimization approaches have been demonstrated in benchmark studies reported in their original publications [@koda2024locating; @koda2024flatbott; @koda2025correlat]. Using a dataset of 121 representative chemical reactions involving typical elements [@asgeirsson2021nudgedel], DMF reduced the computational cost by approximately 70% compared with conventional NEB calculations. FB-ENM was shown to generate more energetically favorable reaction paths than IDPP and to produce chemically plausible initial paths even for complex reactions where IDPP often fails.
+PyDMF provides reusable research software for double-ended transition-state searches and transition-path analysis. It implements DMF and FB-ENM through ASE, making these methods available as documented, tested, and extensible Python components. The underlying methods and their benchmark performance have been described in previous methodological studies [@koda2024locating; @koda2024flatbott; @koda2025correlat].
 
-Beyond these benchmarks, PyDMF has been adopted as a core component in higher-level computational workflows. Notably, it serves as the transition-state search engine in ColabReaction, a web-based application for transition-state searches [@karasawa2025colabrea]. PyDMF is also used as the underlying path-optimization engine in dmf-g16, a Gaussian wrapper that enables reliable double-ended transition-state searches from native Gaussian input files [@koda2026dmfg16]. These integrations demonstrate that PyDMF is not only a methodological implementation but also a practical and reusable software component that supports accessible and reliable transition-state searches.
+The impact of PyDMF is reflected in its adoption by higher-level computational workflows. It serves as the transition-state search engine in ColabReaction, a web-based application for transition-state searches [@karasawa2025colabrea], and has also been incorporated into an end-to-end workflow for automated transition-path analysis of protein reactions [@ohmura2026pdb2]. The GPU support recently added to PyDMF was motivated by this application and has since been incorporated into the main PyDMF codebase through an external contribution. PyDMF is also used as the underlying path-optimization engine in dmf-g16, a Gaussian wrapper that enables double-ended transition-state searches from native Gaussian input files [@koda2026dmfg16].
+
+Together with recent applications of PyDMF and PyDMF-based workflows in computational chemistry studies [@nakano2025high; @kondo2026osim; @dhiman2026porcine], these integrations and contributions show that PyDMF has developed beyond a methodological reference implementation into reusable research infrastructure for transition-state and transition-path searches.
 
 # AI usage disclosure
 
