### 2026-07-12 d09ec9c
add samples/sample_parallel.py
