+++ README_PARALLEL.md
@@ -2,9 +2,17 @@
 
 ## Overview
 
-PyDMF supports parallel energy and force evaluations at the energy evaluation points along the path. As of v1.2.1, PyDMF implements only thread-based parallelization. In practice, PyDMF simply runs multiple ASE calculators concurrently.
+PyDMF supports parallel energy and force evaluations at the energy evaluation points along the path. As of v1.2.1, PyDMF implements only thread-based parallelization. In practice, PyDMF runs multiple ASE calculators concurrently.
 
-Therefore, when constructing the calculators, users should explicitly allocate computational resources as needed and avoid conflicts in output files or working directories.
+Users are therefore responsible for assigning computational resources to each calculator and for avoiding conflicts in output and intermediate files.
+
+A convenient way to configure the calculators is to define a callable named `calc_factory`. The callable receives an image index `i` and returns an ASE calculator configured for that image:
+
+```python
+calc = calc_factory(i)
+```
+
+When `calc_factory` is passed to `DirectMaxFlux`, PyDMF calls it for each image and automatically attaches the returned calculator to that image. This makes it possible to keep node selection, CPU allocation, memory settings, file names, and working-directory settings in one place.
 
 ## Details
 
@@ -12,35 +20,125 @@ Parallel evaluation is enabled by passing `parallel=True` when constructing a `D
 
 The evaluations are parallelized as follows:
 
-- In the initial step, the energy and force calculations for the two endpoint images, `images[0]` and `images[nimages - 1]`, are performed in parallel.
-- In subsequent steps, the energy and force calculations for the internal images, from `images[1]` to `images[nimages - 2]`, are performed in parallel. The number of these images is `nimages - 2`, which is equal to `nmove`.
+* In the initial step, the energy and force calculations for the two endpoint images, `images[0]` and `images[nimages - 1]`, are performed in parallel.
+* In subsequent steps, the energy and force calculations for the internal images, from `images[1]` to `images[nimages - 2]`, are performed in parallel. The number of internal images is `nimages - 2`, which is equal to `nmove`.
+
+The resource assignments for the endpoint images and the internal images may therefore overlap, because these two groups are evaluated at different stages.
+
 
 ## Example: Gaussian calculators
 
-The following example runs Gaussian calculations on `nmove` nodes. Each calculation is launched on a specified node using `mpirun`, and each image uses a separate output directory. This example assumes that the working directory is located on an NFS-mounted filesystem and is visible from all nodes.
+The following simplified example corresponds to `samples/sample_parallel.py`. Although Gaussian is used here as an example, the main point is to illustrate how computational resources can be assigned to individual images and how conflicts in output and intermediate files can be avoided.
+
+The example assumes that the node names and CPU lists have already been assigned to the individual images. Each Gaussian calculation:
+
+* is launched on a specified node using Open MPI,
+* uses a specified list of CPU IDs (on some systems, specifying only nprocshared may be sufficient),
+* writes its files to a separate `imageXX` directory,
+* uses a separate checkpoint file, and
+* receives 1200 MB of Gaussian memory per assigned CPU.
+
[…]
