+++ tests/test_dmf.py
@@ -37,7 +37,7 @@ def test_maxflux():
     from dmf import DirectMaxFlux, interpolate_fbenm
 
     # read react.xyz and prod.xyz
-    ref_images = [read('sample/react.xyz'), read('sample/prod.xyz')]
+    ref_images = [read('samples/react.xyz'), read('samples/prod.xyz')]
 
     # generate initial path by FB-ENM
     mxflx_fbenm = interpolate_fbenm(ref_images,correlated=True)
