### 2026-07-14 9347123
minor changes for JOSS revision
