### 2026-07-14 10b3f38
Bump version to 1.2.2
