### 2026-07-29 c27de76
revise paper/paper.{bib,md}
