+++ paper/paper.bib
@@ -1,6 +1,6 @@
 @article{nakano2025high, 
   year     = {2025}, 
-  title    = {High-Speed Terpene Pathway Analysis via Double-ended Transition State Search Method Combined with {ML} Potentials Uncovers the Intricate Rearrangement Reaction in Spiroalbatene Biosynthesis}, 
+  title    = {High-Speed {Terpene} Pathway Analysis via Double-ended Transition State Search Method Combined with {ML} Potentials Uncovers the Intricate Rearrangement Reaction in {Spiroalbatene} Biosynthesis}, 
   author   = {Nakano, Moe and Karasawa, Masayuki and Ohmura, Takuto and Terada, Tohru and Sato, Hajime}, 
   journal  = {ChemRxiv}, 
   doi      = {10.26434/chemrxiv-2025-md8k6-v2}, 
@@ -39,8 +39,8 @@
 
 @article{koda2026dmfg16,
   year     = {2026},
-  title    = {dmf‐g16: A Gaussian Wrapper for Reliable Double‐Ended Transition‐State Searches With Native Input Formats},
-  author   = {Koda, Shin‐ichi and Saito, Shinji},
+  title    = {{dmf-g16}: A {Gaussian} Wrapper for Reliable Double‐Ended Transition‐State Searches With Native Input Formats},
+  author   = {Koda, Shin-ichi and Saito, Shinji},
   journal  = {Journal of Computational Chemistry},
   doi      = {10.1002/jcc.70378},
   pages    = {e70378},
@@ -49,7 +49,7 @@
 }
 
 @article{koda2025correlat,
-    author = "Koda, Shin-ichi and Saito, Shinji",
+    author = {Koda, Shin-ichi and Saito, Shinji},
     year = "2025",
     title = "Correlated Flat-Bottom Elastic Network Model for Improved Bond Rearrangement in Reaction Paths",
     journal = "Journal of Chemical Theory and Computation",
@@ -60,7 +60,7 @@
 }
 
 @article{koda2024locating,
-    author = "Koda, Shin-ichi and Saito, Shinji",
+    author = {Koda, Shin-ichi and Saito, Shinji},
     year = "2024",
     title = "Locating Transition States by Variational Reaction Path Optimization with an Energy-Derivative-Free Objective Function",
     journal = "Journal of Chemical Theory and Computation",
@@ -71,7 +71,7 @@
 }
 
 @article{koda2024flatbott,
-    author = "Koda, Shin-ichi and Saito, Shinji",
+    author = {Koda, Shin-ichi and Saito, Shinji},
     year = "2024",
     title = "Flat-Bottom Elastic Network Model for Generating Improved Plausible Reaction Paths",
     journal = "Journal of Chemical Theory and Computation",
@@ -139,7 +139,7 @@
 @article{karasawa2025colabrea,
     author = "Karasawa, Masayuki and Leow, Chee Siang and Yajima, Hideaki and Arai, Shuta and Nishizaki, Hiromitsu and Terada, Tohru and Sato, Hajime",
     year = "2025",
-    title = "{ColabReaction}: Accelerating Transition State Searches with Machine Learning Potentials on Google Colaboratory",
+    title = "{ColabReaction}: Accelerating Transition State Searches with Machine Learning Potentials on {Google Colaboratory}",
     journal = "Journal of Chemical Information and Modeling",
     doi = "10.1021/acs.jcim.5c02398",
     pages = "11908--11914",
@@ -149,7 +149,7 @@
 
 @misc{pydmf,
   author = {Koda, Shin-ichi},
-  title = {PyDMF: A Python package for double-ended reaction path/transition state optimization},
[…]
