### 2026-08-12 25e99b3
add citations
