+++ paper/paper.bib
@@ -177,3 +177,157 @@
     number = "1",
     volume = "106"
 }
+% ---- VASP: official general code citations ----
+
+@article{Kresse1993VASP,
+  author  = {Kresse, Georg and Hafner, J{\"u}rgen},
+  title   = {Ab initio molecular dynamics for liquid metals},
+  journal = {Physical Review B},
+  year    = {1993},
+  volume  = {47},
+  number  = {1},
+  pages   = {558--561},
+  doi     = {10.1103/PhysRevB.47.558}
+}
+
+@article{Kresse1994VASP,
+  author  = {Kresse, Georg and Hafner, J{\"u}rgen},
+  title   = {Ab initio molecular-dynamics simulation of the liquid-metal--amorphous-semiconductor transition in germanium},
+  journal = {Physical Review B},
+  year    = {1994},
+  volume  = {49},
+  number  = {20},
+  pages   = {14251--14269},
+  doi     = {10.1103/PhysRevB.49.14251}
+}
+
+@article{Kresse1996VASPCMS,
+  author  = {Kresse, Georg and Furthm{\"u}ller, J{\"u}rgen},
+  title   = {Efficiency of ab-initio total energy calculations for metals and semiconductors using a plane-wave basis set},
+  journal = {Computational Materials Science},
+  year    = {1996},
+  volume  = {6},
+  number  = {1},
+  pages   = {15--50},
+  doi     = {10.1016/0927-0256(96)00008-0}
+}
+
+@article{Kresse1996VASPPRB,
+  author  = {Kresse, Georg and Furthm{\"u}ller, J{\"u}rgen},
+  title   = {Efficient iterative schemes for ab initio total-energy calculations using a plane-wave basis set},
+  journal = {Physical Review B},
+  year    = {1996},
+  volume  = {54},
+  number  = {16},
+  pages   = {11169--11186},
+  doi     = {10.1103/PhysRevB.54.11169}
+}
+
+% ---- Quantum ESPRESSO: official general citations ----
+
+@article{Giannozzi2009QE,
+  author = {Giannozzi, Paolo and Baroni, Stefano and Bonini, Nicola and Calandra, Matteo and Car, Roberto and Cavazzoni, Carlo and Ceresoli, Davide and Chiarotti, Guido L. and Cococcioni, Matteo and Dabo, Ismaila and Dal Corso, Andrea and de Gironcoli, Stefano and Fabris, Stefano and Fratesi, Guido and Gebauer, Ralph and Gerstmann, Uwe and Gougoussis, Christos and Kokalj, Anton and Lazzeri, Michele and Martin-Samos, Layla and Marzari, Nicola and Mauri, Francesco and Mazzarello, Riccardo and Paolini, Stefano and Pasquarello, Alfredo and Paulatto, Lorenzo and Sbraccia, Carlo and Scandolo, Sandro and Sclauzero, Gabriele and Seitsonen, Ari P. and Smogunov, Alexander and Umari, Paolo and Wentzcovitch, Renata M.},
+  title   = {{QUANTUM ESPRESSO}: a modular and open-source software project for quantum simulations of materials},
+  journal = {Journal of Physics: Condensed Matter},
+  year    = {2009},
+  volume  = {21},
+  number  = {39},
+  pages   = {395502},
+  doi     = {10.1088/0953-8984/21/39/395502}
+}
+
+@article{Giannozzi2017QE,
[…]
