+++ paper/paper.md
@@ -43,7 +43,7 @@ PyDMF provides Python implementations of both methods [@pydmf]. Because reaction
 
 # State of the field
 
-As noted above, existing double-ended optimization methods such as NEB are implemented internally in many atomistic simulation programs. The Atomic Simulation Environment (ASE) [@larsen2017theatomi], a Python package that provides a unified interface to many atomistic simulation programs, also provides its own implementations of NEB and several of its variants. In addition, the image-dependent pair potential (IDPP) method [@smidstrup2014improved] for generating initial paths is available in software such as ORCA and ASE.
+As noted above, existing double-ended optimization methods such as NEB are implemented internally in many atomistic simulation programs. The Atomic Simulation Environment (ASE) [@larsen2017theatomi], a Python package that provides a unified interface to many atomistic simulation programs, also provides its own implementations of NEB and several of its variants. In addition, the image-dependent pair potential (IDPP) method [@smidstrup2014improved] for generating initial paths is available in software such as ORCA [@Neese2012ORCA] and ASE.
 
 From a theoretical standpoint, DMF implemented in PyDMF differs fundamentally from the existing approaches. DMF is based on a variational formulation in which reaction-path optimization is expressed as a well-defined minimization problem with an explicit objective function. In contrast to NEB or string methods, which rely on non-variational schemes, DMF can directly leverage general-purpose nonlinear optimization algorithms. In practice, PyDMF employs the state-of-the-art optimizer IPOPT [@wachter2006ontheimp] via its Python interface cyipopt, enabling efficient and robust optimization without introducing method-specific path-update algorithms.
 
@@ -53,7 +53,7 @@ This design choice necessarily introduces a dependency on external optimization 
 
 PyDMF is designed around a clear separation between reaction-path optimization algorithms and energy evaluation backends. Its core design principle is to focus on the implementation of reaction-path optimization methods, while delegating energy and force evaluations to existing atomistic simulation software. ASE provides a well-established abstraction layer that enables this separation, and PyDMF inherits this abstraction by directly interfacing with ASE rather than reimplementing backend-specific functionality.
 
-Through its integration with ASE, PyDMF can access a broad range of electronic-structure and force-field engines for transition-state searches. Well-established packages such as VASP, Quantum ESPRESSO, CP2K, ORCA, Gaussian, GAMESS, LAMMPS, Amber, and GROMACS can be used as backends without any modification to PyDMF. This design choice preserves backend flexibility while allowing PyDMF to integrate smoothly into diverse atomistic modeling workflows. Usage examples and API documentation are available in the project’s GitHub repository [@pydmf].
+Through its integration with ASE, PyDMF can access a broad range of electronic-structure and force-field engines for transition-state searches. Well-established packages such as VASP [@Kresse1993VASP; @Kresse1994VASP; @Kresse1996VASPCMS; @Kresse1996VASPPRB], Quantum ESPRESSO [@Giannozzi2009QE; @Giannozzi2017QE], CP2K [@Kuhne2020CP2K], ORCA [@Neese2012ORCA], Gaussian [@Gaussian16], GAMESS [@Barca2020GAMESS], LAMMPS [@Thompson2022LAMMPS], Amber [@Case2025Amber], and GROMACS [@Abraham2015GROMACS] can be used as backends without any modification to PyDMF. This design choice preserves backend flexibility while allowing PyDMF to integrate smoothly into diverse atomistic modeling workflows. Usage examples and API documentation are available in the project’s GitHub repository [@pydmf].
 
[…]
