### 2026-08-12 5b078d3
modify some bibtex entries
