+++ paper/paper.bib
@@ -14,7 +14,8 @@
   issn     = {1434-193X}, 
   doi      = {10.1002/ejoc.70473}, 
   number   = {20}, 
-  volume   = {29}
+  volume   = {29},
+  pages = {e70473}
 }
 
 @article{ohmura2026pdb2, 
@@ -277,7 +278,7 @@
 
 @misc{Gaussian16,
   author = {Frisch, M. J. and Trucks, G. W. and Schlegel, H. B. and Scuseria, G. E. and Robb, M. A. and Cheeseman, J. R. and Scalmani, G. and Barone, V. and Petersson, G. A. and Nakatsuji, H. and Li, X. and Caricato, M. and Marenich, A. V. and Bloino, J. and Janesko, B. G. and Gomperts, R. and Mennucci, B. and Hratchian, H. P. and Ortiz, J. V. and Izmaylov, A. F. and Sonnenberg, J. L. and Williams-Young, D. and Ding, F. and Lipparini, F. and Egidi, F. and Goings, J. and Peng, B. and Petrone, A. and Henderson, T. and Ranasinghe, D. and Zakrzewski, V. G. and Gao, J. and Rega, N. and Zheng, G. and Liang, W. and Hada, M. and Ehara, M. and Toyota, K. and Fukuda, R. and Hasegawa, J. and Ishida, M. and Nakajima, T. and Honda, Y. and Kitao, O. and Nakai, H. and Vreven, T. and Throssell, K. and Montgomery, Jr., J. A. and Peralta, J. E. and Ogliaro, F. and Bearpark, M. J. and Heyd, J. J. and Brothers, E. N. and Kudin, K. N. and Staroverov, V. N. and Keith, T. A. and Kobayashi, R. and Normand, J. and Raghavachari, K. and Rendell, A. P. and Burant, J. C. and Iyengar, S. S. and Tomasi, J. and Cossi, M. and Millam, J. M. and Klene, M. and Adamo, C. and Cammi, R. and Ochterski, J. W. and Martin, R. L. and Morokuma, K. and Farkas, O. and Foresman, J. B. and Fox, D. J.},
-  title        = {{Gaussian 16}, Revision C.01},
+  title        = {{Gaussian 16}, {Revision C.01}},
   year         = {2016},
   howpublished = {Gaussian, Inc., Wallingford CT}
 }
