### 2026-08-16 1a15f5c
modify citations
