### 2026-08-16 2156ab3
Optimize dmf.torch memory use
