+++ src/dmf/torch/dmf.py
@@ -1,4 +1,5 @@
 import threading
+import warnings
 from abc import ABC, abstractmethod
 from typing import Optional
 
@@ -31,23 +32,63 @@ def _interp1d_torch(xq: torch.Tensor, xp: torch.Tensor, fp: torch.Tensor) -> tor
     return f0 + (f1-f0)*lam
 
 
-def _release_calc_device_cache(calc, *, empty_cache: bool = False):
-    if calc is None:
+def _release_calc_device_cache(calc, seen, errors):
+    """Release each calculator in a mixer graph once."""
+    if calc is None or id(calc) in seen:
         return
+    seen.add(id(calc))
 
     release = getattr(calc, "release_device_cache", None)
     if callable(release):
         try:
-            release(empty_cache=empty_cache)
-        except TypeError:
+            # A single no-arg call lets an implementation TypeError propagate
+            # without being mistaken for signature incompatibility.
             release()
+        except BaseException as exc:
+            errors.append(exc)
 
     mixer = getattr(calc, "mixer", None)
-    if mixer is not None:
-        sub_calcs = getattr(mixer, "calcs", None)
-        if sub_calcs is not None:
-            for sub in sub_calcs:
-                _release_calc_device_cache(sub, empty_cache=False)
+    sub_calcs = getattr(mixer, "calcs", ()) if mixer is not None else ()
+    for sub in sub_calcs:
+        _release_calc_device_cache(sub, seen, errors)
+
+
+def _calc_graph_nodes(calc):
+    """Return calculators reachable through ASE mixer children once each."""
+    ids = set()
+    nodes = []
+    stack = [calc]
+    while stack:
+        current = stack.pop()
+        if current is None or id(current) in ids:
+            continue
+        ids.add(id(current))
+        nodes.append(current)
+        mixer = getattr(current, "mixer", None)
+        stack.extend(getattr(mixer, "calcs", ()) if mixer is not None else ())
+    return nodes
+
+
+def _calc_graph_ids(calc):
+    """Return calculator identities reachable through ASE mixer children."""
+    return {id(node) for node in _calc_graph_nodes(calc)}
+
+
+def _calc_graph_device_cache_entries(calc):
+    """Return unique cache-group owner/device entries for one graph."""
+    entries = {}
+    for node in _calc_graph_nodes(calc):
+        get_key = getattr(node, "_dmf_device_cache_key", None)
+        if not callable(get_key):
+            continue
+        key = get_key()
+        if key is not None:
+            device = torch.device(getattr(node, "device"))
+            bare_cuda = device.type == "cuda" and device.index is None
+            if bare_cuda:
+                device = torch.device("cuda", torch.cuda.current_device())
+            entries.setdefault(id(key), (node, device, bare_cuda))
+    return list(entries.items())
 
 
 class HistoryBase():
@@ -184,6 +225,7 @@ class VariationalPathOpt(ABC, cyipopt.Problem):
 
     n_vel : int, optional
         Number of discretized velocity constraints.
+        Must be at least 1 when specified.
         Default: ``4 * nsegs``.
 
     n_trans : int, optional
@@ -205,6 +247,13 @@ class VariationalPathOpt(ABC, cyipopt.Problem):
     dtype : str or torch.dtype, optional
         Torch floating-point dtype for internal calculations
         (``float32`` or ``float64``). Default: ``float64``.
+
+    keep_history : bool, optional
+        If True, ``intermediate`` records the per-iteration quantities listed in
+        :class:`HistoryBase`. Set to False to skip that record, and the energy
+        interpolation it needs, when the history is not going to be read; the
+        retained forces and structures grow with the iteration count.
+        Default: True.
 
 
     Attributes
[…]
