+++ src/dmf/torch/fbenm.py
@@ -1,3 +1,4 @@
+import threading
 from typing import Optional, Tuple, Union
 
 import numpy as np
@@ -14,6 +15,145 @@ from ._torch_config import _resolve_torch_device, _resolve_torch_dtype
 def _calc_dist_mat(pos, *, device, dtype) -> torch.Tensor:
     pos_t = torch.as_tensor(pos, dtype=dtype, device=device)
     return torch.cdist(pos_t,pos_t)
+
+
+def _actual_device(device) -> torch.device:
+    """Resolve a bare CUDA device to the indexed device used for allocation."""
+    resolved = torch.device(device)
+    if resolved.type == "cuda" and resolved.index is None:
+        return torch.device("cuda", torch.cuda.current_device())
+    return resolved
+
+
+class _SharedConstant(np.ndarray):
+    """Constant shared by every copy of a calculator, so it cannot be written."""
+
+    def __setitem__(self, key, value):
+        if not self.flags.writeable:
+            raise ValueError(
+                "This constant is shared with every copy of this calculator and "
+                "with its device cache, so it is read-only. Build a new "
+                "calculator to use different bounds or wall widths.")
+        super().__setitem__(key, value)
+
+
+def _readonly_view(array: np.ndarray) -> np.ndarray:
+    view = array.view(_SharedConstant)
+    view.setflags(write=False)
+    return view
+
+
+class _FBConstantsOnDevice:
+    """One device-resident copy of the FB-ENM constants, shared by a cache group.
+
+    A cache group is a calculator and the copies made from it: they share these
+    arrays, so one upload serves all of them.  Calculators built independently
+    have their own group, and therefore their own device copy.  Without this the
+    same four N x N matrices were re-uploaded on every force call.  The inverse
+    squared wall widths are stored rather than the widths, since only they are
+    read.
+
+    The device copy is dropped by ``FB_ENM.release_device_cache()``.  The host
+    constants are read-only because copies share both them and this holder;
+    construct a new calculator to use different bounds or wall widths.
+    """
+
+    __slots__ = ("d_min", "d_max", "delta_min", "delta_max", "_lock",
+                 "_key", "_d_min_t", "_d_max_t", "_inv_dmin2_t", "_inv_dmax2_t")
+
+    def __init__(self, d_min, d_max, delta_min, delta_max):
+        self.d_min = d_min
+        self.d_max = d_max
+        self.delta_min = delta_min
+        self.delta_max = delta_max
+        self._lock = threading.Lock()
+        self._key = None
+        self._d_min_t = None
+        self._d_max_t = None
+        self._inv_dmin2_t = None
+        self._inv_dmax2_t = None
+
+    # Device tensors and the lock are runtime state, not part of the value.
+    def __getstate__(self):
+        return {"d_min": self.d_min, "d_max": self.d_max,
+                "delta_min": self.delta_min, "delta_max": self.delta_max}
+
+    def __setstate__(self, state):
+        self.__init__(state["d_min"], state["d_max"],
+                      state["delta_min"], state["delta_max"])
+
+    @torch.no_grad()
+    def get(self, device, dtype):
+        device = _actual_device(device)
+        key = (device, dtype)
+        with self._lock:
+            if self._key != key or self._d_min_t is None:
+                # Drop a previous key before allocating its replacement so a
+                # dtype/device switch cannot transiently retain two 4 x N^2 sets.
[…]
